Fix occasional issue with dropdown rendering
Sometimes when viewing the offline version of an exam, dropdowns take
the full width of the page. My current hypothesis is that this is caused
by a race condition in the exam rendering code. If the stylesheet hasn't
loaded before the exam is rendered, the dropdown is rendered as a block
element since the default styling for <div> elements is applied.  This
causes the rendering code to incorrectly calculate dropdown width to be
100%.

As a fix, inline the content of main.css to offline HTML files directly.
This ensures that the CSS has been applied before the rendering code is
fired, and as a side benefit, it removes potential flashing, since all
critical CSS is now included in the document head.

This increases the size of the HTML files by approximately 41kb (the
current size of main.css), but I consider that to be worth the tradeoff.