Transcode audio & video when creating media versions
We need to transcode .webm and .ogg files to something that most
browsers support. MP3 audio (with libmp3lame) and H264 video (with
libx264) in an MP4 container seem to still offer the best compatibility
across the board, so that's the target format for transcoding.

I decided to use the default encoding parameters ('-q:a 4' for
libmp3lame) and (-preset medium -crf 23 for libx264), since absolute
quality isn't really a high priority for us in the media versions. It's
more important that the media in the exam are of high quality.