generator: allow max-answers in questions
Change the generator API slightly. Code like

const section = {
  questions: [[textAnswer()]]
}

is now written as

const section = {
  questions: [{
    answers: [textAnswer()]
  }]
}

It is also now possible to define the max-answers property of a question
element.

const section = {
  questions: [{
    maxAnswers: 1,
    answers: [textAnswer(), textAnswer()]
  }]
}

To help with the added boilerplate, there is a  new helper function
question() for creating questions.

Like the exam XML itself, the question function is overloaded. There are
three main ways of calling question The simplest way is to provide the
full question structure object.

const section = {
  questions: [question({
    maxAnswers: 1,
    questions: [
      question(…),
      question(…),
    ]
  })]
}

There are also two shorthand forms that are useful when you only want to
provide the answers or child questions of question. In that case, you
can directly provide them as an array.

const section = {
  questions: [
    question([ // Child question shorthand
      question(…),
      question(…),
      question(…),
    ]),
    question([ // Answers shorthand
      textAnswer(),
      textAnswer(),
      textAnswer()
    ])
  ]
}