Fix most of our React's descendant warnings
Right now, we get quite a bit of the following kinds of errors to the
browser console:

   <div> cannot appear as a descendant of <p>

This is caused by invalid HTML in our exam structure, sometimes due to
incorrect XML (that the XML schema doesn't notice) and sometimes by
oversights in our React components.

This commit fixes most of the errors on the React side:

- Make ResponsiveMediaContainer a <span> instead of <figure>. The goal
is to allow <e:image> wherever were allow <img>, like inside <p> tags,
so we need to make sure that <e:image> is phrasing content. The previous
implementation used <figure> for hopes of better semantic compatibility,
but since <figure> is a flow element, it cannot be used inside a <p>.
- Same for <e:attachment>
- Make <e:figure mode="display"> a <span> instead of <div> for the same
reasons.

Fixing all of the errors in the XML side would require us to make the
XML schema stricter. We'll leave that for another day.