Remove extra aria-label from question scores
Normal elements like divs or spans may not have aria-labels if they
don't have the role attribute set, so in practice, this feature was more
trouble than it was worth. VoiceOver didn't read it at all when
shuffling through headings.

In any case, screen reader user should be able to figure what "p" means
from the context, and as a bonus, "p" is shorter.