Remove <Results /> from main-bundle.js
Right now, we ship also the code for viewing exam results in the exam
UMD bundle. This is of course wasteful, and it potentially is also a
security risk, since we are shipping extra code to the client.

As the project grows and we add new features (such as HVP), the pressure
for not including redundant code in the exam bundle also grows.

We tried splitting the bundles before, but it failed since
@digabi/exam-engine-core didn't declare its dependencies correctly, so
trying to import e.g. the Results component resulted in webpack
complaining about missing libraries.

As a fix, declare all runtime dependencies of @digabi/exam-engine-core
in its package.json and generate the UMD bundle from a separate main.ts
entrypoint.

This also improves the tree-shaking experience in downstream projects
that use e.g. the Results component, since code like

{ import Results, parseExam } from '@digabi/exam-engine-core'

will now import the code from an ES module instead of an UMD module, so
tree-shaking should be greatly improved.