Prevent double whitespace from table of contents and hints
Rendering e.g. <p>1.&nbsp; Something</p> causes two space characters
between "1." and "Something". So I suppose it's best not to use NBSP in
these cases at all. This does mean that there may be in theory a newline
between "1." and "Something", but I don't think it's something that we
have to worry about in practice.