Add support for not answering a choice-answer
Add a new (undocumented) `type="no-answer"` attribute to choice-answer,
which when selected, deletes the answer to the question in ktp.

This feature will be used in some matriculation exams to reduce manual
labor - currently we have to clean up the "En vastaa" answers by hand in
exams that contain choice-answers within optional sections.  It also
should be made obsolete when the student is able to select the set of
optional questions to be returned to grading.

Also, as a small bugfix, do not actually delete the answer in
indexedDbExamServerAPI when the value is '' - this matches the semantics
of ktp more accurately, since empty answers are filtered out when
exporting the answers from ktp.