Add an optional assistive-title attribute to e:formula.
This attribute can be used to customize the screen reader representation
of a formula. By default, we hide the SVG image with aria-hidden and
show the LaTeX source of the formula in a sibling screen reader only
element.  With the assistive-title attribute, the screen reader
representation can be customized with e.g. an ASCIIMath or Unicode
version of the formula.