Fix flash on unstyled SVG in offline exams
By default, font-awesome inserts SVG styles to the head element with
JavaScript. This is OK in normal circumstances, but in offline exams,
we've inlined all other CSS to the head element, which causes the
font-awesome styles to be injected only after the JavaScript on the page
has started executing.

As a fix, add the font-awesome CSS to the CSS file ourselves so they are
on the page from the get-go.

An example of the issue: http://yle.fi/plus/abitreenit/2020/syksy/2020-09-17_GE_fi/

Look at the "Vastaa tehtävään 1" icon when reloading the page.