Set lang attributes correctly in foreign-language exams
- Create a wrapper for i18next's useTranslation() hook that wraps a
translated text within a span containing the correct lang attribute, if
necessary.
- Add the proper lang attribute in a few strategic places where we have
localized content within HTML attributes (we can't use the span trick in
those places)
- Add a eslint rule that instructs the programmer to use our custom
translation hook instead of useTranslation()

As a side note, I think the fact that we need to resort to these kinds
of hacks means that either we need to rethink our localization strategy
or perhaps use a different solution than react-i18next. I suppose that
react-i18next isn't _really_ designed for this kind of mixed-language
content that our exams are.