React 17 (#2027)
* Improve redux store type inference

* Replace the last use sites of connect() with hooks

They caused some errors with React 17 typings, so it seems like a fine
time to do this refactoring.

* Upgrade React to 17.x

* Remove react-dom from the UMD bundle

We've passed react-dom in the koe AMD loader since 1st of April 2021, so
it seems like a safe time to remove react-dom from the UMD bundle. This
should enable bundles to work with both React 16 and 17, since the use
site uses passes their own version of React and React-DOM.

* Decrease maximum bundle size to 213kB.

With react-dom gone from the bundle, it is a lot slimmer.