Vapautetaan kokeen rakenteita (#2725)
* Add type attribute for section in exam generator

* Remove sections for hearing-impaired exam by exam-type, not just first section.

When there is a section that is not for hearing-impaired exam, it should have exam-type of "normal" or "normal visually-impaired" defined. It will be removed from hearing-impaired exam with it.

Question numbers starts now from 1 for hearing-impaired exam. They are not in sync with normal exam anymore.

* Same question ids for same questions.

Add question ids before removing questions in wrong language and for other exam-types in order to keep question ids same within different exam versions.
It helps when grading productive questions.

Co-authored-by: Antti Sulanto <anttiv@gmail.com>