Writing mode for rich-text answers (#3115)
Add a writing mode for rich-text answers

Rich-text answers can be opened in full-screen mode that provides an uncluttered view that enables the user to concentrate on that question only. The view consists of the question, textarea and character count, but any images (possibly large) are hidden. The textarea has large paddings and line height, and a better font for writing long answers.