Show exam structure on the left side of the exam (#3127)
Show exam structure on the side of the exam.

The structure element can be used to navigate to different questions. It also highlights currently visible questions, shows which questions have been answered, and displays any errors that there may be regarding answer lengths or the amount of questions answered.

The structure element is also responsive—it gets narrower with the window. The exam structure is also visible in Bertta editor and exam previews.