Move useContext(AnnotationContext) from renderTextNode to AnnotatableText.tsx;
This was causing an error to be printed into console when dropdown option was first selected in the page. After the change, there is no error. The original culprit is the conditional rendering for selectedItem in dropdownanswer. Choice answers do not contain this condition, so there was no error there.